add discussions page in subgraph
